♻️REFACTOR: EventResults 클래스 이름 변경 기존: EventResults -> 변경: ChristmasEventStrategy 및 OrderedMenus 변경 반영.
